requirements.txt
# Dependências do Leitor Ebook
# Instale com: pip install -r requirements.txt

Flask==3.1.0
Flask-Talisman==1.1.0
Flask-Limiter==3.9.4
PyMuPDF==1.24.14
ebooklib==0.18
beautifulsoup4==4.12.3
python-magic==0.4.27
mobi==0.3.3
Werkzeug==3.1.3
